backend/requirements.txt
apscheduler==3.6.3
Django==3.1.2
djangorestframework==3.12.1
gunicorn==20.0.4
psycopg2-binary==2.8.6